pyproject.toml
[tool.poetry]
name = "koppiebot"
version = "0.1.0"
description = "Answers questions based on koplowiczandsons.com"
authors = ["Batman <jordan@koplowicz.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.1.1"
openai = "^1.30.1"
pandas = "^2.2.2"
numpy = "^1.26.4"
python-daemon = "^3.0.1"
discord-py = "^2.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
